#23421 Rename Testcase
Give testcase a meaningful name
